# dynoris/dynoris/swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"Dynoris API"},"basePath":"/","paths":{"/api/Dynoris/CacheItem":{"post":{"tags":["Dynoris"],"operationId":"ApiDynorisCacheItemPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"req","in":"body","required":false,"schema":{"$ref":"#/definitions/CacheItemRequest"}}],"responses":{"200":{"description":"Success"}}}},"/api/Dynoris/CommitItem/{cacheKey}/{updateKey}":{"get":{"tags":["Dynoris"],"operationId":"ApiDynorisCommitItemByCacheKeyByUpdateKeyGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"cacheKey","in":"path","required":true,"type":"string"},{"name":"updateKey","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"Success","schema":{"type":"string"}}}}},"/api/Dynoris/DeleteItem":{"post":{"tags":["Dynoris"],"operationId":"ApiDynorisDeleteItemPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"cacheKey","in":"body","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"Success"}}}},"/api/Dynoris/CacheHash":{"post":{"tags":["Dynoris"],"operationId":"ApiDynorisCacheHashPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"req","in":"body","required":false,"schema":{"$ref":"#/definitions/CacheItemRequest"}}],"responses":{"200":{"description":"Success"}}}},"/api/Dynoris/CacheAsHash":{"post":{"tags":["Dynoris"],"operationId":"ApiDynorisCacheAsHashPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"req","in":"body","required":false,"schema":{"$ref":"#/definitions/CacheItemRequest"}}],"responses":{"200":{"description":"Success"}}}},"/api/ExpiringStamp/Next":{"post":{"tags":["ExpiringStamp"],"operationId":"ApiExpiringStampNextPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"req","in":"body","required":false,"schema":{"$ref":"#/definitions/ExpiringStampRequest"}}],"responses":{"200":{"description":"Success","schema":{"type":"array","items":{"type":"string"}}}}}},"/api/ExpiringStamp/CommitItem":{"post":{"tags":["ExpiringStamp"],"operationId":"ApiExpiringStampCommitItemPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"req","in":"body","required":false,"schema":{"$ref":"#/definitions/CommitItemRequest"}}],"responses":{"200":{"description":"Success"}}}}},"definitions":{"CacheItemRequest":{"type":"object","properties":{"cacheKey":{"type":"string"},"table":{"type":"string"},"storeKey":{"type":"array","items":{"$ref":"#/definitions/ValueTuple[String,String]"}},"indexName":{"type":"string"},"hashKey":{"type":"string"}}},"ValueTuple[String,String]":{"type":"object","properties":{"item1":{"type":"string"},"item2":{"type":"string"}}},"ExpiringStampRequest":{"type":"object","properties":{"table":{"type":"string"},"index":{"type":"string"},"storeKey":{"type":"array","items":{"$ref":"#/definitions/ValueTuple[String,String]"}},"stampKey":{"$ref":"#/definitions/ValueTuple[String,String]"}}},"CommitItemRequest":{"type":"object","properties":{"table":{"type":"string"},"storeKey":{"type":"array","items":{"$ref":"#/definitions/ValueTuple[String,String]"}},"itemJson":{"type":"string"}}}},"securityDefinitions":{}}